Extract controller type to bypass bootstrap
